Rename hook-up-sticky-nav to hookUpStickyNav
