+++ src/components/NewFilter/NewFilter.jsx
@@ -1,7 +1,7 @@
 const React = require('react')
 const Filter = require('../../models/filter')
 const FilterHelp = require('../FilterHelp')
-const hookUpStickyNav = require('../hook-up-sticky-nav')
+const hookUpStickyNav = require('../hookUpStickyNav')
 
 class NewFilter extends React.Component {
   constructor() {
