+++ src/components/TaskList/TaskList.jsx
@@ -2,7 +2,7 @@ const React = require('react')
 const { connect } = require('react-redux')
 
 const TaskListItem = require('../TaskListItem')
-const hookUpStickyNav = require('../hook-up-sticky-nav')
+const hookUpStickyNav = require('../hookUpStickyNav')
 const Filters = require('../../models/filters')
 const LastFilter = require('../../models/last-filter')
