+++ src/components/hookUpStickyNav/package.json
@@ -2,5 +2,5 @@
   "name": "hookUpStickyNav",
   "version": "0.0.0",
   "private": true,
-  "main": "./hook-up-sticky-nav.jsx"
+  "main": "./hookUpStickyNav.jsx"
 }
